refactor: move `Keybindings` to its own file
